requirements.txt
watchdog
google-api-python-client
google-auth-oauthlib
google-auth-httplib2